requirements.txt
Django == 3.1
psycopg2
django-tinymce
requests
Pillow
